+++ compiler-core/src/type_/tests/errors.rs
@@ -3091,18 +3091,8 @@ fn double_assignment_in_bit_array() {
 }
 
 #[test]
-fn negative_size_value() {
-    assert_error!("<<1:size(-1)>>");
-}
-
-#[test]
 fn negative_size_pattern() {
     assert_error!("let assert <<1:size(-1)>> = <<>>");
-}
-
-#[test]
-fn zero_size_value() {
-    assert_error!("<<1:size(0)>>");
 }
 
 #[test]