+++ CHANGELOG.md
@@ -173,8 +173,8 @@
   Erlang target.
   ([Surya Rose](https://github.com/GearsDatapacks))
 
-- The compiler will now reject bit array segments whose constant size is zero or
-  negative. Previously a zero or negative sized segment would crash the
+- The compiler will now reject bit array segment patterns whose constant size is
+  zero or negative. Previously a zero or negative sized segment would crash the
   compiler.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))